Separate desktop nav and add a collision hook
